Interface is upgraded
