+++ mobile/Mercatus/app/src/main/java/com/bounswe/mercatus/Adapters/CommentAdapter.kt
@@ -11,7 +11,7 @@ import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.RecyclerView
 import com.bounswe.mercatus.API.ApiInterface
 import com.bounswe.mercatus.API.RetrofitInstance
-import com.bounswe.mercatus.Fragments.ShowProfileActivity
+import com.bounswe.mercatus.Fragments.User.ShowProfileActivity
 import com.bounswe.mercatus.Models.CommentEditBody
 import com.bounswe.mercatus.Models.CommentShowBody
 import com.bounswe.mercatus.Models.UserRes
