+++ mobile/Mercatus/app/src/main/java/com/bounswe/mercatus/Adapters/CustomAdapter.kt
@@ -6,7 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.bounswe.mercatus.Fragments.ShowProfileActivity
+import com.bounswe.mercatus.Fragments.User.ShowProfileActivity
 import com.bounswe.mercatus.Models.SearchShow
 import com.bounswe.mercatus.R
 import kotlinx.android.synthetic.main.item_layout.view.*
