+++ mobile/Mercatus/app/src/main/java/com/bounswe/mercatus/Fragments/Articles/ShowArticleActivity.kt
@@ -13,7 +13,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bounswe.mercatus.API.ApiInterface
 import com.bounswe.mercatus.API.RetrofitInstance
 import com.bounswe.mercatus.Adapters.CommentAdapter
-import com.bounswe.mercatus.Fragments.ShowProfileActivity
+import com.bounswe.mercatus.Fragments.User.ShowProfileActivity
 import com.bounswe.mercatus.Models.*
 import com.bounswe.mercatus.R
 import kotlinx.android.synthetic.main.activity_show_article.*
