+++ mobile/Mercatus/app/src/main/java/com/bounswe/mercatus/Fragments/Entrance/BasicFragment.kt
@@ -1,4 +1,4 @@
-package com.bounswe.mercatus.Fragments
+package com.bounswe.mercatus.Fragments.Entrance
 
 import android.content.Intent
 import android.os.Bundle
@@ -7,17 +7,17 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.bounswe.mercatus.R
-import kotlinx.android.synthetic.main.fragment_trader.*
+import kotlinx.android.synthetic.main.fragment_basic.*
 
-class TraderFragment : Fragment() {
+class BasicFragment : Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_trader, container, false)
+        return inflater.inflate(R.layout.fragment_basic, container, false)
     }
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        btnTrader.setOnClickListener {
-            val intent = Intent(activity, TraderUserActivity::class.java)
+        btnBasic.setOnClickListener {
+            val intent = Intent(activity, BasicUserActivity::class.java)
             startActivity(intent) }
         activity?.overridePendingTransition(
             R.anim.slide_in_right,
