+++ mobile/Mercatus/app/src/main/java/com/bounswe/mercatus/Fragments/Entrance/BasicUserActivity.kt
@@ -1,4 +1,4 @@
-package com.bounswe.mercatus.Fragments
+package com.bounswe.mercatus.Fragments.Entrance
 
 import android.app.DatePickerDialog
 import android.content.Context
@@ -10,36 +10,36 @@ import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.bounswe.mercatus.API.ApiInterface
 import com.bounswe.mercatus.API.RetrofitInstance
+import com.bounswe.mercatus.Fragments.SearchActivity
 import com.bounswe.mercatus.Models.SignInBody
 import com.bounswe.mercatus.Models.SignInRes
 import com.bounswe.mercatus.Models.UserBody
 import com.bounswe.mercatus.Models.UserRes
 import com.bounswe.mercatus.R
-import kotlinx.android.synthetic.main.activity_trader_user.*
+import kotlinx.android.synthetic.main.activity_basic_user.*
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 import java.net.ConnectException
 import java.util.*
 
-class TraderUserActivity : AppCompatActivity() {
+class BasicUserActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_trader_user)
+        setContentView(R.layout.activity_basic_user)
 
         val c = Calendar.getInstance()
         pickDate(c)
+
         floatingActionButton.setOnClickListener{
             val name = editName.text.toString()
             val surname = editSurname.text.toString()
             val email = editMail.text.toString()
             val password = editPassword.text.toString()
-            val iban = editIban.text.toString()
-            val id  = editID.text.toString()
 
             val date = dateTv.text.toString()
-            if (isValidForm(email, name, surname, password, iban, id, date)) {
+            if (isValidForm(email, name, surname, password, date)) {
                 signup(date, email, name, surname, password)
             }
         }
@@ -52,14 +52,14 @@ class TraderUserActivity : AppCompatActivity() {
             override fun onFailure(call: Call<SignInRes>, t: Throwable) {
                 if(t.cause is ConnectException){
                     Toast.makeText(
-                        this@TraderUserActivity,
+                        this@BasicUserActivity,
                         "Check your connection!",
                         Toast.LENGTH_SHORT
                     ).show()
                 }
                 else{
                     Toast.makeText(
-                        this@TraderUserActivity,
+                        this@BasicUserActivity,
                         "Something bad happened!",
                         Toast.LENGTH_SHORT
                     ).show()
@@ -70,9 +70,9 @@ class TraderUserActivity : AppCompatActivity() {
                     editor.putString("token", response.body()?.token)
                     editor.apply()
 
-                    Toast.makeText(this@TraderUserActivity, "Login success!.", Toast.LENGTH_SHORT).show()
+                    Toast.makeText(this@BasicUserActivity, "Login success!.", Toast.LENGTH_SHORT).show()
 
-                    val intent = Intent(this@TraderUserActivity, SearchActivity::class.java)
+                    val intent = Intent(this@BasicUserActivity, SearchActivity::class.java)
                     startActivity(intent)
                     overridePendingTransition(
                         R.anim.slide_in_right,
@@ -80,30 +80,28 @@ class TraderUserActivity : AppCompatActivity() {
                     )
 
                 } else {
-                    Toast.makeText(this@TraderUserActivity, "Login failed!", Toast.LENGTH_SHORT).show()
+                    Toast.makeText(this@BasicUserActivity, "Login failed!", Toast.LENGTH_SHORT).show()
                 }
             }
         })
     }
-
-    private fun isValidForm(email: String, name: String, surname: String,
-                            password: String, iban: String, id: String, date:String):Boolean{
+    private fun isValidForm(email: String, name: String, surname: String, password: String, date: String):Boolean{
 
         var isValid = true
 
         if (name.isEmpty()){
-            layPassword.isErrorEnabled = true
-            layPassword.error = "Name cannot be empty!"
+            layName.isErrorEnabled = true
+            layName.error = "Name cannot be empty!"
             isValid = false
         }else{
-            layPassword.isErrorEnabled = false
+            layName.isErrorEnabled = false
         }
         if (surname.isEmpty()){
-            layPassword.isErrorEnabled = true
-            layPassword.error = "Surname cannot be empty!"
+            laySurname.isErrorEnabled = true
+            laySurname.error = "Surname cannot be empty!"
             isValid = false
         }else{
-            layPassword.isErrorEnabled = false
+            laySurname.isErrorEnabled = false
         }
         if (!email.isValidEmail()){
             layMail.isErrorEnabled = true
@@ -116,20 +114,6 @@ class TraderUserActivity : AppCompatActivity() {
             layPassword.isErrorEnabled = true
             layPassword.error = "Password cannot be empty!"
             isValid = false
-        }else{
-            layPassword.isErrorEnabled = false
-        }
-        if (iban.length > 24 || iban.length < 24){
-            layIban.isErrorEnabled = true
-            layIban.error = "Iban is not valid!"
-            isValid = false
-        }else{
-            layIban.isErrorEnabled = false
-        }
-        if (id.length > 11 || id.length < 11){
-            layID.isErrorEnabled = true
-            layID.error = "TC id is not valid!"
-            isValid = false
         }
         else if(password.length <6){
             layPassword.isErrorEnabled = true
@@ -137,7 +121,7 @@ class TraderUserActivity : AppCompatActivity() {
             isValid = false
         }
         else{
-            layID.isErrorEnabled = false
+            layPassword.isErrorEnabled = false
         }
         if (date == "Date"){
             dateTv.error = "Date cannot be empty!"
@@ -158,14 +142,14 @@ class TraderUserActivity : AppCompatActivity() {
             override fun onFailure(call: Call<UserRes>, t: Throwable) {
                 if(t.cause is ConnectException){
                     Toast.makeText(
-                        this@TraderUserActivity,
+                        this@BasicUserActivity,
                         "Check your connection!",
                         Toast.LENGTH_SHORT
                     ).show()
                 }
                 else{
                     Toast.makeText(
-                        this@TraderUserActivity,
+                        this@BasicUserActivity,
                         "Something bad happened!",
                         Toast.LENGTH_SHORT
                     ).show()
@@ -173,15 +157,16 @@ class TraderUserActivity : AppCompatActivity() {
             }
             override fun onResponse(call: Call<UserRes>, response: Response<UserRes>) {
                 if (response.code() == 201) {
-                    Toast.makeText(this@TraderUserActivity, "Registration success!", Toast.LENGTH_SHORT)
+                    Toast.makeText(this@BasicUserActivity, "Registration success!", Toast.LENGTH_SHORT)
                         .show()
                     val sharedPreferences = getSharedPreferences("TOKEN_INFO", Context.MODE_PRIVATE)
                     val editor = sharedPreferences.edit()
 
                     //Since we need token, we need to do signin
                     signin(email, password, editor)
-                } else {
-                    Toast.makeText(this@TraderUserActivity, "Registration failed!", Toast.LENGTH_SHORT)
+                }
+                else  {
+                    Toast.makeText(this@BasicUserActivity, "Registration failed", Toast.LENGTH_SHORT)
                         .show()
                 }
             }
@@ -193,7 +178,7 @@ class TraderUserActivity : AppCompatActivity() {
         val day = c.get(Calendar.DAY_OF_MONTH)
 
         dataPick.setOnClickListener{
-            val dpd = DatePickerDialog(this, DatePickerDialog.OnDateSetListener{ view, mYear, Mmonth, mDayOfMonth ->
+            val dpd = DatePickerDialog(this, DatePickerDialog.OnDateSetListener{view, mYear, Mmonth, mDayOfMonth ->
                 dateTv.setText("" + mYear + "-" + Mmonth + "-" + mDayOfMonth)
             }, year, month, day)
             dpd.show()
