+++ mobile/Mercatus/app/src/main/java/com/bounswe/mercatus/Fragments/User/ShowProfileActivity.kt
@@ -1,4 +1,4 @@
-package com.bounswe.mercatus.Fragments
+package com.bounswe.mercatus.Fragments.User
 
 import android.content.Context
 import android.content.Intent
