+++ mobile/Mercatus/app/src/main/res/layout/activity_basic_user.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragments.BasicUserActivity"
+    tools:context=".Fragments.Entrance.BasicUserActivity"
     android:background="@color/colorPrimary">
 
     <FrameLayout
