+++ mobile/Mercatus/app/src/main/res/layout/activity_followers.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragments.FollowersActivity">
+    tools:context=".Fragments.User.FollowersActivity">
 
 
     <androidx.recyclerview.widget.RecyclerView
