+++ mobile/Mercatus/app/src/main/res/layout/activity_followings.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragments.FollowingsActivity">
+    tools:context=".Fragments.User.FollowingsActivity">
 
 
     <androidx.recyclerview.widget.RecyclerView
