+++ mobile/Mercatus/app/src/main/res/layout/activity_trader_user.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragments.TraderUserActivity"
+    tools:context=".Fragments.Entrance.TraderUserActivity"
     android:background="@color/colorPrimary">
 
     <FrameLayout
