+++ mobile/Mercatus/app/src/main/res/layout/activity_trading.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragments.TradingActivity">
+    tools:context=".Fragments.Entrance.TradingActivity">
 
     <ir.farshid_roohi.linegraph.LineChart
         android:id="@+id/lineChart"
