+++ mobile/Mercatus/app/src/main/res/layout/nav_header_main.xml
@@ -18,7 +18,7 @@
         android:layout_height="60dp"
         android:contentDescription="@string/nav_header_desc"
         android:paddingTop="@dimen/nav_header_vertical_spacing"
-        app:srcCompat="@drawable/icon" />
+        app:srcCompat="@drawable/mer_no_background" />
 
     <TextView
         android:layout_width="match_parent"
